interface
coneccion lista en passos

+++ app/src/main/java/com/example/pavilion/proyectomoviles/Services/DeficientModel.java
@@ -6,12 +6,12 @@ package com.example.pavilion.proyectomoviles.Services;
 public class DeficientModel {
     private String name;
     private String description;
-    private int R;
-    private int G;
-    private int B;
-    private int A;
+    private float R;
+    private float G;
+    private float B;
+    private float A;
 
-    public DeficientModel(String name, String description, int r, int g, int b, int a) {
+    public DeficientModel(String name, String description, float r, float g, float b, float a) {
         this.name = name;
         this.description = description;
         R = r;
@@ -36,35 +36,35 @@ public class DeficientModel {
         this.description = description;
     }
 
-    public int getR() {
+    public float getR() {
         return R;
     }
 
-    public void setR(int r) {
+    public void setR(float r) {
         R = r;
     }
 
-    public int getG() {
+    public float getG() {
         return G;
     }
 
-    public void setG(int g) {
+    public void setG(float g) {
         G = g;
     }
 
-    public int getB() {
+    public float getB() {
         return B;
     }
 
-    public void setB(int b) {
+    public void setB(float b) {
         B = b;
     }
 
-    public int getA() {
+    public float getA() {
         return A;
     }
 
-    public void setA(int a) {
+    public void setA(float a) {
         A = a;
     }
 }
